refactor(backend): standardize file naming and update imports\n\n- Rename middleware and utils files to dot-delimited convention (internal.api.key.middleware.js, validate.input.middleware.js, server.util.js)\n- Update all imports/routes to new filenames\n- Enforce internal API key under /api\n- Adjust tests to new paths (all green)\n\nchore(docs): update sessions API contract\n- Document start/complete/list/summary endpoints and FE snippets\n- Clarify timestamps and error shape\n\nchore(gitignore): broaden monorepo ignores
