+++ apps/backend/src/routes/v1/authentication.routes.js
@@ -43,7 +43,7 @@ import {
   forgotPassword,
   resetPassword,
 } from '../../controllers/auth.controller.js';
-import { validateSignup } from '../../middleware/validateInput.js';
+import { validateSignup } from '../../middleware/validate.input.middleware.js';
 
 const router = Router();
